Exception handling - deleteById

+++ src/main/java/com/danielpsilva/course/services/UserService.java
@@ -4,9 +4,11 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.danielpsilva.course.entities.User;
+import com.danielpsilva.course.exceptions.DatabaseException;
 import com.danielpsilva.course.exceptions.ResourceNotFoundException;
 import com.danielpsilva.course.repositories.UserRepository;
 
@@ -30,7 +32,15 @@ public class UserService {
 	}
 
 	public void deleteById(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (userRepository.existsById(id)) {
+				userRepository.deleteById(id);
+			} else {
+				throw new ResourceNotFoundException(id);
+			}
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 
 	public User update(Long id, User obj) {
